+++ src/lib.rs
@@ -17,3 +17,4 @@ pub use pidsk_controller::PidController;
 pub use pidsk_controller::{Pid, Pidf32, Pidf64};
 pub use pidsk_controller::{PidErrorf32, PidErrorf64, PidErrors};
 pub use pidsk_controller::{PidGains, PidGainsf32, PidGainsf64};
+pub use pidsk_controller::{PidLimits, PidLimitsf32, PidLimitsf64};